requirements.txt
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-model-utils==4.1.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==2.10
lxml==4.6.3
multitasking==0.0.9
numpy==1.21.0
oauthlib==3.1.1
pandas==1.2.5
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
yfinance==0.1.63
